Исправление предупреждений xrGame

+++ xrGame/ai/monsters/control_animation.cpp
@@ -111,7 +111,7 @@ void CControlAnimation::play_part(SAnimationPart& part, PlayCallback callback)
 	// initialize synchronization of prev and current animation
 	float pos = -1.f;
 	if (part.blend && !part.blend->stop_at_end)
-		pos = fmod(part.blend->timeCurrent, part.blend->timeTotal) / part.blend->timeTotal;
+		pos = fmodf(part.blend->timeCurrent, part.blend->timeTotal) / part.blend->timeTotal;
 #ifdef DEBUG
 		// CKinematicsAnimated * K = m_object->Visual()->dcast_PKinematicsAnimated();
 		// Msg				("%6d Playing animation : %s , %s , Object %s",Device.dwTimeGlobal,
