+++ xrGame/stalker_animation_pair.cpp
@@ -115,7 +115,7 @@ void CStalkerAnimationPair::play(CKinematicsAnimated* skeleton_animated, PlayCal
 		{
 			VERIFY(!m_blend || !fis_zero(m_blend->timeTotal));
 			if (m_step_dependence && m_blend)
-				pos = fmod(m_blend->timeCurrent, m_blend->timeTotal) / m_blend->timeTotal;
+				pos = fmodf(m_blend->timeCurrent, m_blend->timeTotal) / m_blend->timeTotal;
 		}
 
 		m_blend = skeleton_animated->PlayCycle(animation(), TRUE, callback, object);
